image-recognition-v2/requirements_cpu.txt
torch==1.8.0+cpu
torchvision==0.9.0+cpu
torchaudio==0.8.0+cpu
easyocr
asyncio
transformers[torch]
fastapi
uvicorn
python-dotenv